requirements.txt
fastapi
uvicorn
ultralytics
Pillow
python-multipart
streamlit